feat: export commitlint function
