+++ packages/dev-lib/src/index.ts
@@ -1,1 +1,2 @@
 export * from './config.js'
+export * from './commitlint.util.js'
